chore: move for AzureLib
