+++ src/main/java/yesman/epicfight/compat/azurelib/AzureLibArmorCompat.java
@@ -3,7 +3,7 @@ package yesman.epicfight.compat.azurelib;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.api.distmarker.OnlyIn;
 import net.neoforged.bus.api.IEventBus;
-import yesman.epicfight.api.client.model.transformer.AzureArmorTransformer;
+import yesman.epicfight.compat.azurelib.client.AzureArmorTransformer;
 import yesman.epicfight.api.client.model.transformer.HumanoidModelBaker;
 import yesman.epicfight.compat.ICompatModule;
