+++ src/main/java/yesman/epicfight/compat/azurelib/AzureLibCompat.java
@@ -6,7 +6,7 @@ import net.minecraft.world.entity.LivingEntity;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.api.distmarker.OnlyIn;
 import net.neoforged.bus.api.IEventBus;
-import yesman.epicfight.api.client.model.transformer.AzureModelTransformer;
+import yesman.epicfight.compat.azurelib.client.AzureModelTransformer;
 import yesman.epicfight.api.client.model.transformer.HumanoidModelBaker;
 import yesman.epicfight.client.ClientEngine;
 import yesman.epicfight.client.events.engine.RenderEngine;
